feat: refactor repository services with base class abstraction
- Create BaseRepository abstract class with shared functionality
- Refactor GitHubRepository to extend BaseRepository
- Move common interfaces and methods to base class
- Update GitLab project converter specification to completed status
- Establish foundation for validation and packaging tasks
